Update patch set 2
Patch Set 2: Code-Review-1

I don't think this is the right way to fix the validate role atm.
The switch to fedora image is a temporary fix, I submit a review for switching to cirros: https://review.gerrithub.io/#/c/277197/
with a specific kernel and initramfs for cirros image to prevent the issue at the boot.

Patch-set: 2
Label: Code-Review=-1